deeplearning/rag_pipeline/app/requirement.txt
fastapi==0.115.0
uvicorn==0.22.0
python-dotenv==1.0.0
mistralai==0.0.7
